Pizza Page, Confirmation, Cart is Full, Error
Created the Confirmation and Cart is full popups. Currently there is no pizza preview on the pizza page; we may need to discuss how we do that. Also you can scroll on the pizza page both vertically and horizontally, which I will try to fix later. There is an error popup for the pizza page but it's still work in progress.